feat(Poolboy): add poolboy dependency and configure to run
The Poolboy dependency has been added to the Mixfile and the supervisor
has been configured to run Poolboy as a child process.

Poolboy has also been added as a dependency to the application list to
prevent a failure when building an OTP release.

A new worker has been created that acts as the GenServer for Poolboy to
communicate with when running the pooled processes. This worker
currently has no purpose as it doesn't run any process or return
anything useful.