feat(Poolboy): run `Squarer.square` process via Poolboy worker
The Poolboy worker has now been updated to get the square of the given
number and return the result in the GenServer reply.

The main application has also been updated to call a Poolboy worker via
a transaction with the given number. Another method has been added that
tests the basic implementation of the Poolboy setup.