feat(Poolboy): implement parallel processing for `Squarer.square`
Another function has been added that will spawn processes for each
number in a given range and push each number in the list to the Poolboy
worker for processing.

As these spawned processes won't return anything back to the main
application, the Worker has been updated to print out the current
background task that it is running.

There is an issue with this commit that Poolboy will timeout if it
cannot allocated a worker within the default time of 5 seconds and as
soon as it times out it will no longer take on any new jobs.