docs(README): describe implementing parallel processes for workers
Spawning new processes that call a Poolboy transaction is described
along with the limitation that spawned processes do not return any data
to the original calling function.

A warning has been added to highlight the issue with the commit if large
ranges are provided as Poolboy will timeout by default after 5 seconds
if no workers are available to handle the process.