feat(Poolboy): ensure Poolboy waits infinitely for a worker
To prevent Poolboy from timing out whilst waiting for a worker, the
`:infinity` atom is passed as the `TIMEOUT` argument in the Poolboy
transaction function call. This will ensure Poolboy waits forever for a
worker to become available.

To highlight the issue more, a sleep has been added to the worker to
simulate slow running processes.